// apps/ts/packages/domain/src/dataset/types.ts
export interface StockInfo {
  code: string;
  companyName: string;
  companyNameEnglish: string;
  marketCode: string;
  marketName: string;
  sector17Code: string;
  sector17Name: string;
  sector33Code: string;
  sector33Name: string;
  scaleCategory: string;
  listedDate: Date;
}